pyproject.toml
[project]
name = "image-to-minecraft"
version = "0.1.0"
authors = [
  { name="Nikita Gamygin", email="nikita.gamygin@gmail.com" },
]
description = "Takes an image and recreates it from a set of given blocks"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
license = "MIT"
dependencies = [
    "pillow>=12.0.0"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
